# laravel/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartSprouts API",
        "description": "API documentation for SmartSprouts backend",
        "contact": {
            "email": "support@smartsprouts.local"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get details for a specific game",
                "description": "Returns the full details for a game by its ID.",
                "operationId": "getGameDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with game details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get authenticated user profile",
                "description": "Returns user profile data with aggregated gameplay statistics.",
                "operationId": "getUserProfile",
                "responses": {
                    "200": {
                        "description": "Successful response with profile data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProfileStats": {
                "title": "Profile Statistics",
                "properties": {
                    "totalScore": {
                        "type": "integer",
                        "example": 1250
                    },
                    "totalLevels": {
                        "type": "integer",
                        "example": 45
                    },
                    "completedLevels": {
                        "type": "integer",
                        "example": 12
                    },
                    "correctAnswersPercentage": {
                        "type": "number",
                        "format": "float",
                        "example": 88.5
                    }
                },
                "type": "object"
            },
            "Profile": {
                "title": "User Profile",
                "required": [
                    "name",
                    "email",
                    "stats"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "stats": {
                        "$ref": "#/components/schemas/ProfileStats"
                    }
                },
                "type": "object"
            },
            "Game": {
                "title": "Game",
                "required": [
                    "id",
                    "key",
                    "icon_url",
                    "is_active",
                    "title",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "key": {
                        "type": "string",
                        "example": "find_the_wrong"
                    },
                    "title": {
                        "description": "Full title of the game.",
                        "type": "string",
                        "example": "Find The Wrong"
                    },
                    "description": {
                        "description": "Short description of the game.",
                        "type": "string",
                        "example": "Find the incorrect statement among the options."
                    },
                    "icon_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/icons/game1.png"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}